Fix: change to getSelfStream (thanks to Chuck)
In the delta reduction it would not recognize speedThroughWater because PGN 128259 contains both speedThoughWater and speedThroughWaterReferenceType. The attached file replaced all the if's and looking at EVERY delta on the server. I changed it to app.streambundle.getSelfStream for just the paths needed to make it all work. I also fixed the wind index look up. I tested it against my polar table with SKSim at known angles and wind speeds and the data matched the expected values. I will update again after a race this coming weekend. Thanks... Chuck